fix: massive plugin log payloads
It would seem if a queued web request to our plugin logs failed, it would continually try to send more triage logs to the same non-working endpoint until the server crashed. also, in some cases we assumed the response would be json, and on some errors we might get an html document. now we handle this appropriately.